Add tox, which adds linting and doc building to testing

+++ autoapi/mappers/dotnet.py
@@ -7,6 +7,7 @@ import shutil
 import yaml
 from sphinx.util.osutil import ensuredir
 from sphinx.util.console import darkgreen, bold
+from sphinx.errors import ExtensionError
 
 from .base import PythonMapperBase, SphinxMapperBase
 
@@ -23,28 +24,32 @@ class DotNetSphinxMapper(SphinxMapperBase):
 
     top_namespaces = {}
 
-    def load(self, patterns, dir, ignore=[]):
+    def load(self, patterns, dir, ignore=None, **kwargs):
         '''
         Load objects from the filesystem into the ``paths`` dictionary.
 
         '''
+        raise_error = kwargs.get('raise_error', True)
         all_files = list(self.find_files(patterns=patterns, dir=dir, ignore=ignore))
-        try:
-            command = ['docfx', 'metadata', '--raw', '--force']
-            command.extend(all_files)
-            proc = subprocess.Popen(
-                command,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                env=dict((key, os.environ[key])
-                         for key in ['PATH', 'DNX_PATH', 'HOME']
-                         if key in os.environ),
-            )
-            _, error_output = proc.communicate()
-            if error_output:
-                self.app.warn(error_output)
-        except (OSError, subprocess.CalledProcessError) as e:
-            self.app.warn('Error generating metadata: {0}'.format(e))
+        if all_files:
+            try:
+                command = ['docfx', 'metadata', '--raw', '--force']
+                command.extend(all_files)
+                proc = subprocess.Popen(
+                    command,
+                    stdout=subprocess.PIPE,
+                    stderr=subprocess.PIPE,
+                    env=dict((key, os.environ[key])
+                             for key in ['PATH', 'DNX_PATH', 'HOME']
+                             if key in os.environ),
+                )
+                _, error_output = proc.communicate()
+                if error_output:
+                    self.app.warn(error_output)
+            except (OSError, subprocess.CalledProcessError) as e:
+                self.app.warn('Error generating metadata: {0}'.format(e))
+                if raise_error:
+                    raise ExtensionError('Failure in docfx while generating AutoAPI output.')
         # We now have yaml files
         for xdoc_path in self.find_files(patterns=['*.yml'], dir='_api_', ignore=ignore):
             data = self.read_file(path=xdoc_path)
@@ -115,7 +120,7 @@ class DotNetSphinxMapper(SphinxMapperBase):
         :param obj: Instance of a .NET object
         '''
         if obj.top_level_object:
-            if type(obj) == DotNetNamespace:
+            if isinstance(obj, DotNetNamespace):
                 self.namespaces[obj.name] = obj
         self.objects[obj.id] = obj
 
@@ -138,7 +143,7 @@ class DotNetSphinxMapper(SphinxMapperBase):
             namespace = obj.top_namespace
             if namespace is not None:
                 ns_obj = self.top_namespaces.get(namespace)
-                if ns_obj is None or type(ns_obj) != DotNetNamespace:
+                if ns_obj is None or not isinstance(ns_obj, DotNetNamespace):
                     for ns_obj in self.create_class({'uid': namespace,
                                                      'type': 'namespace'}):
                         self.top_namespaces[ns_obj.id] = ns_obj
@@ -189,7 +194,8 @@ class DotNetSphinxMapper(SphinxMapperBase):
     def build_finished(app, exception):
         if app.verbosity > 1:
             app.info(bold('[AutoAPI] ') + darkgreen('Cleaning generated .yml files'))
-        shutil.rmtree('_api_')
+        if os.path.exists('_api_'):
+            shutil.rmtree('_api_')
 
 
 class DotNetPythonMapper(PythonMapperBase):
